Successfully stored user data in Database after verifying
So what we're doing here is that - at first we've set up a Router for the /register page in auth.js file and then added some logic to verify whether a user already exist or not. - Then nodemon app.js is fired and we sent the user data from POSTMAN. In POSTMAN we can see the confirm and error messages that we've written in auth.js. - If the user is new and has filled all the details then the submitted data will be stored in Atlas database!